requirements.txt
anyio==3.7.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
fastapi==0.97.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
MarkupSafe==2.1.3
pydantic==1.10.9
PyJWT==2.7.0
PyMySQL==1.0.3
python-multipart==0.0.6
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.3.4